RFIDapi_Python/requirements.txt
aiohttp
aiohttp-cors
websockets
psutil
mysql-connector-python
paho-mqtt
pydantic
opencv-contrib-python
